\name{Lab4LinearRegression-package}
\alias{Lab4LinearRegression-package}
\alias{Lab4LinearRegression}
\docType{package}
\title{
\packageTitle{Lab4LinearRegression}
}
\description{
\packageDescription{Lab4LinearRegression}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab4LinearRegression}
\packageIndices{Lab4LinearRegression}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Lab4LinearRegression}

Maintainer: \packageMaintainer{Lab4LinearRegression}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
