% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-Lab4LinearRegression.R
\docType{package}
\name{Lab4LinearRegression}
\alias{Lab4LinearRegression}
\title{Lab4LinearRegression: A package for creating a linear regression model}
\description{
The Lab4LinearRegression package provides many important functions using an RC class
}
