% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear regression model}
\value{
nothing
}
\description{
returns a linear regression model for formula and data inputs

This is a function that builds a linear regression model using ordinary least squares method and RC class.
A reference class for linreg
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula with dependent numeric variable on left and independent numeric variable on right.}

\item{\code{data}}{A dataset}

\item{\code{formula}}{as character}

\item{\code{data}}{as data.frame}

\item{\code{call}}{as vector}

\item{\code{coefficients_est}}{as matrix}

\item{\code{fitted_vals}}{as matrix}

\item{\code{residuals_est}}{as matrix}

\item{\code{degrees_of_freedom}}{as numeric}

\item{\code{res_var_est}}{as matrix}

\item{\code{v_coeff}}{as vector}

\item{\code{find_t_vals}}{as numeric}

\item{\code{p_vals}}{as vector}

\item{\code{std_res}}{as matrix}
}}


