% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Building a linear regression model using ordinary least squares method}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{object is a formula with dependent numeric variable on left and independent numeric variable on right.}

\item{data}{A numeric matrix.}
}
\value{
Statistics for linear regression model.
}
\description{
Building a linear regression model using ordinary least squares method
}
